Rename d_exercise_id to the exercise_id in tables in_workout_item and in_warmup_workout_item.

+++ src/main/java/com/osomapps/pt/admin/user/UserWarmupWorkoutItemResponseDTO.java
@@ -13,7 +13,7 @@ import lombok.experimental.Accessors;
 @Setter
 class UserWarmupWorkoutItemResponseDTO {
     Long id;
-    Long exercise_id;
+    Integer exercise_id;
     String exercise_name;
     Integer speed;
     Integer incline;
